Add files via upload
The Wether-Finder webpage is a simple weather application that allows users to enter a city name and get the current weather information for that location. The application fetches data from a weather API and displays details such as temperature, weather conditions, and humidity. It is designed with a basic user interface using HTML, CSS, and JavaScript. You can check out the live site here.